fix : Template 수정
- recipe 오타 수정
- 필요없는 SVG 삭제
- MainIcon 컴포넌트 통합
